+++ json_parse.c
@@ -1413,6 +1413,7 @@ decode_json_string(char const *ptr, size_t len, size_t mlen, size_t *retlen)
 		}
 		xa = 0;
 		xb = 0;
+
 		/*
 		 * we check for a second \uxxxx first, in case it is a surrogate
 		 * pair
@@ -1435,6 +1436,18 @@ decode_json_string(char const *ptr, size_t len, size_t mlen, size_t *retlen)
 		     * no possible surrogate pair found so proceed like there
 		     * was not another \uxxxx
 		     */
+
+		    /*
+		     * first we check if the range is in the non-character range as
+		     * the spec recommend setting it to the replacement character.
+		     * *sigh*
+		     */
+		    if (is_unicode_noncharacter(xa)) {
+			dbg(DBG_MED, "%s: %X is non-character, setting to replacement character: U+%X", __func__,
+				UNICODE_REPLACEMENT_CHAR);
+			xa = UNICODE_REPLACEMENT_CHAR;
+		    }
+
 		    bytes = utf8encode(utf8, xa);
 		    if (bytes < 0) {
 			/* error - clear allocated length and free buffer */
@@ -1709,12 +1722,19 @@ json_decode(char const *ptr, size_t len, size_t *retlen)
 		} else if (scanned == 1 || (scanned == 2 && surrogates_to_unicode(xa, xb) < 0)) {
 		    surrogate = xa;
 		    bytes = utf8len(ptr + i, surrogate);
-		    if (bytes <= 0) {
+		    if (bytes <= 0 && bytes != UNICODE_NOT_CHARACTER) {
 			if (retlen != NULL) {
 			    *retlen = 0;
 			}
 			/* utf8len() already warns */
 			return NULL;
+		    } else if (bytes == UNICODE_NOT_CHARACTER) {
+			/*
+			 * according to the spec recommendation, a character in
+			 * the non character range should be set to the
+			 * replacement character which is U+FFFD and is 3 bytes.
+			 */
+			bytes = 3;
 		    }
 		    dbg(DBG_VVHIGH, "UTF-8 bytes: %ju", (uintmax_t)bytes);
 		    mlen += bytes;